Add streaming API

+++ CChromaEditorLibrary/CChromaEditorLibrary.cpp
@@ -765,7 +765,7 @@ BOOL CMainViewDlg::OnInitDialog()
 	GetControlSetKeyCombo()->SetCurSel(0);
 
 	// setup mouse chars
-	for (int led = EChromaSDKMouseLED::ML_SCROLLWHEEL; led <= EChromaSDKMouseLED::ML_RIGHT_SIDE7; ++led)
+	for (int led = (int)EChromaSDKMouseLED::ML_SCROLLWHEEL; led <= (int)EChromaSDKMouseLED::ML_RIGHT_SIDE7; ++led)
 	{
 		const char* strLed = ChromaSDKPlugin::GetInstance()->GetMouseChar((EChromaSDKMouseLED)led);
 		GetControlSetLEDCombo()->AddString(CString(strLed));
